requirements/base.txt
# web
Django==1.11.12
djangorestframework==3.7.7
  coreapi==2.3.3
  drf-yasg[validation]==1.3.1
  django-filter==1.0.4
  django-cors-headers==2.1.0
  drf-extensions==0.3.1
django-crispy-forms==1.7.0
django-extensions==1.8.1
django-selectable==1.1.0
django-webpack-loader==0.5.0
django-markdown-deux==1.0.5

# tox21
jupyter==1.0.0
#numpy==1.13.0
numpy==1.13.3
pandas==0.20.2
  xlrd==1.0.0
  openpyxl==2.4.7
psycopg2==2.7.4
#psycopg2>=2.8

# docs
mkdocs==0.17.2

# enricher
# rdkit==2017.09.3  # install via conda not pip
#cython==0.28.2
#git+http://github.com/rdkit/django-rdkit@30cac1f61a1d401a5020a5e410a648fd4b2bfbda

# fiveday
#bmds==0.10.0
#jinja2==2.9.5
#pdfkit==0.6.1
matplotlib==2.2.2
#beautifulsoup4==4.5.3
scipy==0.19.0
scikit-learn==0.18.1
docxtpl==0.3.5
#jsonschema==2.5.1
seaborn==0.7.1

# CEBS
#cx-Oracle==6.0.2

# neurotox
plotly==2.0.12

# job_runner
#celery==4.2.0
#kombu==4.2.1
#pyRserve==0.9.1

